fix 3.0

+++ models/totaltender_model.go
@@ -5,7 +5,7 @@ package models
 type Totaltender struct {
 	Name      string  `json:"name,omitempty" validate:"required"`
 	Total     float32 `json:"total,omitempty" validate:"required"`
-	Totalpagu float32 `json:"totalpagu,omitempty" validate:"required"`
+	Totalpagu float64 `json:"totalpagu,omitempty" validate:"required"`
 	Pdn       float32 `json:"pdn,omitempty" validate:"required"`
 	Idpagu    string  `json:"idpagu,omitempty" validate:"required"`
 }
